tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.enable.x.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.enable.x</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.enable.x</h1>
<p class="desc">Shadows are drawn if shadowOffsetX is set</p>


<script>
var t = async_test("Shadows are drawn if shadowOffsetX is set");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.globalCompositeOperation = 'destination-atop';
  ctx.shadowColor = '#0f0';
  ctx.shadowOffsetX = 0.1;
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
